Generate coordinates of small circles

Generate coordinates of small circles

Usage

sc_shape(
  x = "random",
  r = NULL,
  r.ex = NULL,
  r.rad = NULL,
  r.deg = NULL,
  n = NULL,
  breaks = 100,
  radius = 6371.007,
  origin = c(0, 0, 0),
  output = "polar",
  sf.type = "polygon",
  sf.wrap.dateline = TRUE,
  drop = TRUE
)

Arguments

x

Numeric matrix or character string. Coordinates of the centers of the small circles

r

The surface radius of the small circle expressed as great circle distance of the small circle's points from the center of the small circles.

r.ex

The extrinsic radius of the small circles, i.e. the radius of the circle in the plane of the small circle.

r.rad

The angle of the small circle in radians.

r.deg

The angle of the small circle in degrees.

n

Integer or NULL. only used if "x= random"

breaks

Integer. The number of points to create from the small circle.

radius

The radius of the sphere, defaults to the authalic radius of Earth.

origin

The center of the sphere (don't touch this, unless you know what you are doing!).

output

Output structure for the function. The value output="polar" will return the polar (longitude-latitude) coordinates of the small circles in an array.

drop

If there is a single small circle to generate, should its array wrapper be dropped? The setting output="cartesian" will return the 3D cartesian coordinates of the small circles in an array. The option output="sf" will return an sfc geometry collection.

Value

Either a numeric array or and sfc geometry collection.